chore.更新 C、Python 集成示例
